app/test_ocr_main/requirements.txt
# Requirements for TEST_OCR project
craft_text_detector
opencv-python
Pillow
torch
torchvision
transformers
